first pass: separate some core events
