+++ core/dbt/common/events/eventmgr.py
@@ -1,6 +1,7 @@
 import os
 import traceback
 from typing import Callable, List, Optional, Protocol, Tuple
+from types import ModuleType
 from uuid import uuid4
 
 from dbt.common.events.base_types import BaseEvent, EventLevel, msg_from_base_event, EventMsg
@@ -38,6 +39,9 @@ class EventManager:
         )
         self.loggers.append(logger)
 
+    def add_event_protos(self, event_proto_module: ModuleType) -> None:
+        BaseEvent._PROTO_TYPES_MODULES.update(event_proto_module.__dict__)
+
     def flush(self) -> None:
         for logger in self.loggers:
             logger.flush()
@@ -52,6 +56,9 @@ class IEventManager(Protocol):
         ...
 
     def add_logger(self, config: LoggerConfig) -> None:
+        ...
+
+    def add_event_protos(self, event_proto_module: ModuleType) -> None:
         ...
