+++ core/dbt/common/events/functions.py
@@ -12,6 +12,7 @@ import json
 import os
 import sys
 from typing import Callable, Dict, List, Optional, TextIO, Union
+from types import ModuleType
 import uuid
 from google.protobuf.json_format import MessageToDict
 
@@ -34,10 +35,15 @@ def make_log_dir_if_missing(log_path: Union[Path, str]) -> None:
     log_path.mkdir(parents=True, exist_ok=True)
 
 
-def setup_event_logger(flags, callbacks: List[Callable[[EventMsg], None]] = []) -> None:
+# TODO: make None default
+def setup_event_logger(
+    flags, callbacks: List[Callable[[EventMsg], None]] = [], event_modules: List[ModuleType] = []
+) -> None:
     cleanup_event_logger()
     make_log_dir_if_missing(flags.LOG_PATH)
     EVENT_MANAGER.callbacks = callbacks.copy()
+    for event_module in event_modules:
+        EVENT_MANAGER.add_event_protos(event_module)
 
     if flags.LOG_LEVEL != "none":
         line_format = _line_format_from_str(flags.LOG_FORMAT, LineFormat.PlainText)
