+++ core/dbt/common/events/types.proto
@@ -287,134 +287,6 @@ message ProjectCreatedMsg {
     ProjectCreated data = 2;
 }
 
-// D - Deprecation
-
-// D001
-message PackageRedirectDeprecation {
-    string old_name = 1;
-    string new_name = 2;
-}
-
-message PackageRedirectDeprecationMsg {
-    EventInfo info = 1;
-    PackageRedirectDeprecation data = 2;
-}
-
-// D002
-message PackageInstallPathDeprecation {
-}
-
-message PackageInstallPathDeprecationMsg {
-    EventInfo info = 1;
-    PackageInstallPathDeprecation data = 2;
-}
-
-// D003
-message ConfigSourcePathDeprecation {
-    string deprecated_path = 1;
-    string exp_path = 2;
-}
-
-message ConfigSourcePathDeprecationMsg {
-    EventInfo info = 1;
-    ConfigSourcePathDeprecation data = 2;
-}
-
-// D004
-message ConfigDataPathDeprecation {
-    string deprecated_path = 1;
-    string exp_path = 2;
-}
-
-message ConfigDataPathDeprecationMsg {
-    EventInfo info = 1;
-    ConfigDataPathDeprecation data = 2;
-}
-
-// D005
-message AdapterDeprecationWarning {
-    string old_name = 1;
-    string new_name = 2;
-}
-
-message AdapterDeprecationWarningMsg {
-    EventInfo info = 1;
-    AdapterDeprecationWarning data = 2;
-}
-
-// D006
-message MetricAttributesRenamed {
-    string metric_name = 1;
-}
-
-message MetricAttributesRenamedMsg {
-    EventInfo info = 1;
-    MetricAttributesRenamed data = 2;
-}
-
-// D007
-message ExposureNameDeprecation {
-    string exposure = 1;
-}
-
-message ExposureNameDeprecationMsg {
-    EventInfo info = 1;
-    ExposureNameDeprecation data = 2;
-}
-
-// D008
-message InternalDeprecation {
-    string name = 1;
-    string reason = 2;
-    string suggested_action = 3;
-    string version = 4;
-}
-
-message InternalDeprecationMsg {
-    EventInfo info = 1;
-    InternalDeprecation data = 2;
-}
-
-// D009
-message EnvironmentVariableRenamed {
-    string old_name = 1;
-    string new_name = 2;
-}
-
-message EnvironmentVariableRenamedMsg {
-    EventInfo info = 1;
-    EnvironmentVariableRenamed data = 2;
-}
-
-// D010
-message ConfigLogPathDeprecation {
-    string deprecated_path = 1;
-}
-
-message ConfigLogPathDeprecationMsg {
-    EventInfo info = 1;
-    ConfigLogPathDeprecation data = 2;
-}
-
-// D011
-message ConfigTargetPathDeprecation {
-    string deprecated_path = 1;
-}
-
-message ConfigTargetPathDeprecationMsg {
-    EventInfo info = 1;
-    ConfigTargetPathDeprecation data = 2;
-}
-
-// D012
-message CollectFreshnessReturnSignature {
-}
-
-message CollectFreshnessReturnSignatureMsg {
-    EventInfo info = 1;
-    CollectFreshnessReturnSignature data = 2;
-}
-
 // E - DB Adapter
 
 // E001
@@ -1194,18 +1066,6 @@ message UnpinnedRefNewVersionAvailable {
 message UnpinnedRefNewVersionAvailableMsg {
     EventInfo info = 1;
     UnpinnedRefNewVersionAvailable data = 2;
-}
-
-// I065
-message DeprecatedModel {
-    string model_name = 1;
-    string model_version = 2;
-    string deprecation_date = 3;
-}
-
-message DeprecatedModelMsg {
-    EventInfo info = 1;
-    DeprecatedModel data = 2;
 }
 
 // I066
