+++ core/dbt/events/core_types.proto
@@ -1,0 +1,160 @@
+syntax = "proto3";
+
+package proto_types;
+
+import "google/protobuf/timestamp.proto";
+import "google/protobuf/struct.proto";
+
+// Common event info
+message CoreEventInfo {
+    string name = 1;
+    string code = 2;
+    string msg = 3;
+    string level = 4;
+    string invocation_id = 5;
+    int32 pid = 6;
+    string thread = 7;
+    google.protobuf.Timestamp ts = 8;
+    map<string, string> extra = 9;
+    string category = 10;
+}
+
+// D - Deprecation
+
+// D001
+message PackageRedirectDeprecation {
+    string old_name = 1;
+    string new_name = 2;
+}
+
+message PackageRedirectDeprecationMsg {
+    CoreEventInfo info = 1;
+    PackageRedirectDeprecation data = 2;
+}
+
+// D002
+message PackageInstallPathDeprecation {
+}
+
+message PackageInstallPathDeprecationMsg {
+    CoreEventInfo info = 1;
+    PackageInstallPathDeprecation data = 2;
+}
+
+// D003
+message ConfigSourcePathDeprecation {
+    string deprecated_path = 1;
+    string exp_path = 2;
+}
+
+message ConfigSourcePathDeprecationMsg {
+    CoreEventInfo info = 1;
+    ConfigSourcePathDeprecation data = 2;
+}
+
+// D004
+message ConfigDataPathDeprecation {
+    string deprecated_path = 1;
+    string exp_path = 2;
+}
+
+message ConfigDataPathDeprecationMsg {
+    CoreEventInfo info = 1;
+    ConfigDataPathDeprecation data = 2;
+}
+
+// D005
+message AdapterDeprecationWarning {
+    string old_name = 1;
+    string new_name = 2;
+}
+
+message AdapterDeprecationWarningMsg {
+    CoreEventInfo info = 1;
+    AdapterDeprecationWarning data = 2;
+}
+
+// D006
+message MetricAttributesRenamed {
+    string metric_name = 1;
+}
+
+message MetricAttributesRenamedMsg {
+    CoreEventInfo info = 1;
+    MetricAttributesRenamed data = 2;
+}
+
+// D007
+message ExposureNameDeprecation {
+    string exposure = 1;
+}
+
+message ExposureNameDeprecationMsg {
+    CoreEventInfo info = 1;
+    ExposureNameDeprecation data = 2;
+}
+
+// D008
+message InternalDeprecation {
+    string name = 1;
+    string reason = 2;
+    string suggested_action = 3;
+    string version = 4;
+}
+
+message InternalDeprecationMsg {
+    CoreEventInfo info = 1;
+    InternalDeprecation data = 2;
+}
+
+// D009
+message EnvironmentVariableRenamed {
+    string old_name = 1;
+    string new_name = 2;
+}
+
+message EnvironmentVariableRenamedMsg {
+    CoreEventInfo info = 1;
+    EnvironmentVariableRenamed data = 2;
+}
+
+// D010
+message ConfigLogPathDeprecation {
+    string deprecated_path = 1;
+}
+
+message ConfigLogPathDeprecationMsg {
+    CoreEventInfo info = 1;
+    ConfigLogPathDeprecation data = 2;
+}
+
+// D011
+message ConfigTargetPathDeprecation {
+    string deprecated_path = 1;
+}
+
+message ConfigTargetPathDeprecationMsg {
+    CoreEventInfo info = 1;
+    ConfigTargetPathDeprecation data = 2;
+}
+
+// D012
+message CollectFreshnessReturnSignature {
+}
+
+message CollectFreshnessReturnSignatureMsg {
+    CoreEventInfo info = 1;
+    CollectFreshnessReturnSignature data = 2;
+}
+
+// I065
+message DeprecatedModel {
+    string model_name = 1;
+    string model_version = 2;
+    string deprecation_date = 3;
+}
+
+message DeprecatedModelMsg {
+    CoreEventInfo info = 1;
+    DeprecatedModel data = 2;
+}
